requirements.txt
fastapi

uvicorn[standard]

pymongo

celery[redis]

python-dotenv

redis

python-multipart

python-dotenv

moviepy

google-cloud-storage